requirements.txt
requests==2.24.0
terminaltables==3.1.0
python-dotenv==0.14.0